Bugfix: Crash on 64 bit system.
git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@1836 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/svxlink/modules/metarinfo/ModuleMetarInfo.cpp
@@ -685,7 +685,7 @@ int ModuleMetarInfo::onDataReceived(TcpConnection *con, void *buf, int count)
    bool endflag = false;
    bool nceiling = false;
    float temp_view = 0;
-   unsigned int found;
+   size_t found;
    int metartoken;
 
    char *metarinput = static_cast<char *>(buf);
